+++ test/unit/fuzz-tools.test.js
@@ -144,6 +144,7 @@ import * as frailtyV143 from '../../lib/frailty-v143.js';
 import * as orthoV144 from '../../lib/ortho-v144.js';
 import * as orthoV145 from '../../lib/ortho-v145.js';
 import * as spineV146 from '../../lib/spine-v146.js';
+import * as rheumV147 from '../../lib/rheum-v147.js';
 import * as growthLmsData from '../../lib/growth-lms-data.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
@@ -234,6 +235,7 @@ const MODULES = {
   'ortho-v144.js': orthoV144,
   'ortho-v145.js': orthoV145,
   'spine-v146.js': spineV146,
+  'rheum-v147.js': rheumV147,
   'growth-lms-data.js': growthLmsData,
   'ems-v149.js': emsV149,
 };